refactor(agents): extract run_command action execution from controller actions god file
Follow-up to the field_access extraction (#9387). Move the run_command
action-execution cluster out of controller_service/actions.rs (still a
god file): execute_run_command_action + run_controller_command_with_timeout,
the cfg-gated process-group configure/kill/signal helpers, IO dir /
json-file helpers, artifact capture/persist, and required-artifact
validation -> run_command.rs.

The cluster is contiguous (execute_run_command_action ..
artifact_refs_from_command_result). Two of its fns are used elsewhere in
actions.rs (execute_run_command_action from the action dispatcher,
sanitize_loop_action_id from the dispatch-identity path) — both marked
pub(super) and imported back. The cluster also reuses a few actions.rs
helpers (RUN_COMMAND_DEFAULT_TIMEOUT_SECONDS, CappedCommandOutput +
read/collect_capped_command_output), now pub(super). Follows the existing
sibling-module convention (#![allow(unused_imports)] + use super::*).

Behavior-preserving: actions.rs 1859 -> 1533 lines, run_command.rs 344.
All 81 agent_task_controller_service tests pass single-threaded. (The
full parallel suite shows pre-existing with_isolated_home shared-state
flakiness unrelated to this move — verified: tests pass in isolation,
single-threaded, and scoped; the same races exist on clean main.)
